Rename kubefnord to kubefed2
